+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 1.1.0 - Unreleased
+
+### Added
+
+- Added getImage() function to Image as shortcut to getDriver()-\>getImage().
+- Added close() function to Image as shortcut to setHandle(null).
+- Added more image param override options for each driver image format.
+
+### Fixed
+
+- Fixed PNG compression level option in Imagick driver.
+
 ## 1.0.1 - 2023-07-15
 
 ### Fixed